+++ shared/CalculatorService.java
@@ -4,11 +4,11 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 
 public interface CalculatorService extends Remote {
-    double add(double a, double b) throws RemoteException;
+    int add(int a, int b) throws RemoteException;
 
-    double subtract(double a, double b) throws RemoteException;
+    int subtract(int a, int b) throws RemoteException;
 
-    double multiply(double a, double b) throws RemoteException;
+    int multiply(int a, int b) throws RemoteException;
 
-    double divide(double a, double b) throws RemoteException;
+    int divide(int a, int b) throws RemoteException;
 }
